project3-code-agent-langchain/requirements.txt
# LangChain core
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# OpenAI SDK for LLM API calls
openai>=1.0.0

# Environment variable management
python-dotenv>=1.0.0

# Numerical computing
numpy>=1.0.0

# Testing framework
pytest>=7.0.0
